requirements.txt
pyrogram==2.0.106
pyromod==1.5
tgcrypto
aiohttp
aiofiles
requests
hachoir
cloudscraper
pycryptodome
pytz
motor
pyrofork
PyJWT
aiogram
beautifulsoup4
httpx
python-telegram-bot==20.6
tqdm
colorama
termcolor
dotenv

